refactor: Move logic for popup matcher to separate class
